refactoring of keys
